restructure admin page
