+++ src/components/Admin/Product.jsx
@@ -1,0 +1,4 @@
+export default function Product({sneaker}){
+    return <div>{sneaker.name}</div>
+    
+}
