+++ src/pages/App/App.js
@@ -8,6 +8,7 @@ import LoginForm from '../../components/Account/LoginForm/LoginForm';
 import SignUpForm from '../../components/Account/SignUpForm/SignUpForm';
 import ProductPage from '../ProductPage/ProductPage';
 import AdminPage from '../AdminPage/AdminPage';
+import AddInventoryPage from '../AdminPage/AddInventoryPage';
 
 //helper
 import { getUser } from '../../utilities/services/users';
@@ -40,6 +41,7 @@ export default function App() {
               <Route path="/createaccount" element={<SignUpForm setUser={setUser}/>} />
               <Route path="/" element={<ProductPage sneakers={sneakers} SetSneakers={SetSneakers} />} />
               <Route path="/admin" element={<AdminPage user={user} sneakers={sneakers}/>} />
+              <Route path="/admin/addinventory" element={<AddInventoryPage user={user}/>} />
             </Routes>
             <Footer user={user} setUser={setUser}/>
     </main>
