+++ src/utilities/services/sneaker.js
@@ -1,8 +1,4 @@
 import * as adminAPI from '../api/sneakers'
 
 
-// export async function getSneaker(){
-//     const sneaker = await adminAPI.checkSneaker()
-//     return sneaker
-//   }
